project writeup for game dev
This is the project writeup for my game development class from when I had submitted this project. It provides a basic idea of what the game is mechanically, and lists out various "bounties" (things I could do to earn points for a better grade) that I was able to complete. It also lists some credits for assets from other people that I used. The YouTube link doesn't work, and some information may be a little irrelevant / outdated.